# Backend/CookBook/CookBookLibrary/Models/RecipeIngredient.cs
using System;
using System.Collections.Generic;

namespace CookBookLibrary.Models
{
    public partial class RecipeIngredient
    {
        public int RecipeId { get; set; }
        public int IngredientId { get; set; }
        public string Measurement { get; set; } = null!;

        public virtual Ingredient Ingredient { get; set; } = null!;
        public virtual Recipe Recipe { get; set; } = null!;
    }
}