# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CookBookWebAPI",
    "version": "1.0"
  },
  "paths": {
    "/api/ingredients": {
      "get": {
        "tags": [
          "Ingredient"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Ingredient/{ingredientId}": {
      "get": {
        "tags": [
          "Ingredient"
        ],
        "parameters": [
          {
            "name": "ingredientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Ingredient"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ingredient"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ingredient"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Ingredient"
        ],
        "parameters": [
          {
            "name": "ingredientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreationDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Ingredient"
        ],
        "parameters": [
          {
            "name": "ingredientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "Ingredient"
        ],
        "parameters": [
          {
            "name": "ingredientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Ingredient": {
      "post": {
        "tags": [
          "Ingredient"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreationDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/recipes": {
      "get": {
        "tags": [
          "Recipe"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Recipe/{recipeId}": {
      "get": {
        "tags": [
          "Recipe"
        ],
        "parameters": [
          {
            "name": "recipeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Recipe"
        ],
        "parameters": [
          {
            "name": "recipeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreationDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Recipe"
        ],
        "parameters": [
          {
            "name": "recipeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/recipes/cuisine": {
      "get": {
        "tags": [
          "Recipe"
        ],
        "parameters": [
          {
            "name": "cuisine",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Recipe": {
      "post": {
        "tags": [
          "Recipe"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreationDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCreationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Recipe/{recipeId}/{ingredientId}": {
      "patch": {
        "tags": [
          "Recipe"
        ],
        "parameters": [
          {
            "name": "recipeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ingredientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Ingredient": {
        "type": "object",
        "properties": {
          "ingredientId": {
            "type": "integer",
            "format": "int32"
          },
          "ingredientName": {
            "type": "string",
            "nullable": true
          },
          "recipeIngredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeIngredient"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IngredientCreationDto": {
        "type": "object",
        "properties": {
          "ingredientName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Operation": {
        "type": "object",
        "properties": {
          "operationType": {
            "$ref": "#/components/schemas/OperationType"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "op": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OperationType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "Recipe": {
        "type": "object",
        "properties": {
          "recipeId": {
            "type": "integer",
            "format": "int32"
          },
          "recipeName": {
            "type": "string",
            "nullable": true
          },
          "cuisine": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "instruction": {
            "type": "string",
            "nullable": true
          },
          "prepTime": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "string",
            "nullable": true
          },
          "prepVideoUrl": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "complexity": {
            "type": "string",
            "nullable": true
          },
          "recipeIngredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeIngredient"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RecipeCreationDto": {
        "type": "object",
        "properties": {
          "recipeName": {
            "type": "string",
            "nullable": true
          },
          "cuisine": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "instruction": {
            "type": "string",
            "nullable": true
          },
          "prepTime": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "string",
            "nullable": true
          },
          "prepVideoUrl": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "complexity": {
            "type": "string",
            "nullable": true
          },
          "numberOfIngredients": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "recipeIngredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeIngredientCreationDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RecipeIngredient": {
        "type": "object",
        "properties": {
          "recipeId": {
            "type": "integer",
            "format": "int32"
          },
          "ingredientId": {
            "type": "integer",
            "format": "int32"
          },
          "measurement": {
            "type": "string",
            "nullable": true
          },
          "ingredient": {
            "$ref": "#/components/schemas/Ingredient"
          },
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          }
        },
        "additionalProperties": false
      },
      "RecipeIngredientCreationDto": {
        "type": "object",
        "properties": {
          "ingredientId": {
            "type": "integer",
            "format": "int32"
          },
          "measurement": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}